RewirePipelineCommandHandler
